requirements.txt
streamlit
opencv-python
pytesseract
Pillow
numpy
python-dotenv
openai